requirements.txt
fastapi
uvicorn[standard]
openai
langchain
langchain_community
together
pypdf
pyJWT 
passlib
pydantic[email]